fetch STS endpoint from properties
Even when providing `withRegion`, the `AWSSecurityTokenServiceClientBuilder` points the client to us-east-1. This fails in cases where you are using a VPC endpoint and access to the Internet is limited.

This change looks up the property `aws.sts.endpoint` and uses that or `https://sts.amazonaws.com` to pass to `.withEndpointConfiguration`

Closes #58